requirements.txt
# Core dependencies for PDF ingestion and question generation

# Google Generative AI for Gemini API (LLM and embeddings)
google-generativeai>=0.3.0

# Qdrant client for vector database operations
qdrant-client>=1.7.0

# AWS SDK for S3 image storage
boto3>=1.34.0

# Type checking and validation
typing-extensions>=4.8.0

# Optional: for better development experience
python-dotenv>=1.0.0  # For managing environment variables
